important update
I updated the logging process in the main dlt scripts. I plan to remove the 'classify' submodule since it's the copy of the main script despite the operator is not involved. Previous testing finds that it costs more time using the default identity operator than without the operator. In this update it decides to use the operator when the operator is not None. This will save the time. The free adversarial training is modified a lot. It will not depend on the input argument 'perturb' to determine the perturbation method.